reorganising the filestructure
