+++ src/js/events/checkLoginStatus.js
